Move the dynamo bridge to the _dynamo repo (#7909)
