+++ test/dynamo/test_dynamo_aliasing.py
@@ -4,8 +4,7 @@ import torch
 import torch_xla
 import torch_xla.core.xla_model as xm
 import torch_xla.debug.metrics as met
-import torch_xla.experimental.dynamo_set_buffer_donor
-from torch_xla.core.dynamo_bridge import alias_with_buffer_donor_config
+from torch_xla._dynamo.dynamo_bridge import alias_with_buffer_donor_config
 
 
 class TestBufferDonationUtil(unittest.TestCase):
